+++ sdk/playwrighttesting/arm-playwrighttesting/CHANGELOG.md
@@ -1,18 +1,51 @@
 # Release History
-
-## 1.0.0-beta.3 (Unreleased)
-
+    
+## 1.0.0-beta.3 (2024-08-29)
+Compared with version 1.0.0-beta.2
+    
 ### Features Added
+
+  - Added operation group AccountQuotas
+  - Added operation Accounts.checkNameAvailability
+  - Added Interface AccountFreeTrialProperties
+  - Added Interface AccountQuota
+  - Added Interface AccountQuotaListResult
+  - Added Interface AccountQuotaProperties
+  - Added Interface AccountQuotasGetOptionalParams
+  - Added Interface AccountQuotasListByAccountNextOptionalParams
+  - Added Interface AccountQuotasListByAccountOptionalParams
+  - Added Interface AccountsCheckNameAvailabilityOptionalParams
+  - Added Interface CheckNameAvailabilityRequest
+  - Added Interface CheckNameAvailabilityResponse
+  - Added Type Alias AccountQuotasGetResponse
+  - Added Type Alias AccountQuotasListByAccountNextResponse
+  - Added Type Alias AccountQuotasListByAccountResponse
+  - Added Type Alias AccountsCheckNameAvailabilityResponse
+  - Added Type Alias AccountsDeleteResponse
+  - Added Type Alias CheckNameAvailabilityReason
+  - Added Type Alias OfferingType
+  - Interface AccountProperties has a new optional parameter localAuth
+  - Interface AccountUpdateProperties has a new optional parameter localAuth
+  - Interface QuotaProperties has a new optional parameter offeringType
+  - Added Enum KnownCheckNameAvailabilityReason
+  - Added Enum KnownOfferingType
+  - Enum KnownFreeTrialState has a new value NotEligible
+  - Enum KnownFreeTrialState has a new value NotRegistered
+  - Enum KnownProvisioningState has a new value Creating
+  - Enum KnownQuotaNames has a new value Reporting
 
 ### Breaking Changes
 
-### Bugs Fixed
-
-### Other Changes
-
+  - Interface FreeTrialProperties no longer has parameter allocatedValue
+  - Interface FreeTrialProperties no longer has parameter createdAt
+  - Interface FreeTrialProperties no longer has parameter expiryAt
+  - Interface FreeTrialProperties no longer has parameter percentageUsed
+  - Interface FreeTrialProperties no longer has parameter usedValue
+    
+    
 ## 1.0.0-beta.2 (2024-01-31)
     
-**Features**
+### Features Added
 
   - Added Interface AccountProperties
   - Added Interface AccountUpdateProperties
@@ -21,7 +54,7 @@
   - Interface AccountUpdate has a new optional parameter properties
   - Interface Quota has a new optional parameter properties
 
-**Breaking Changes**
+### Breaking Changes
 
   - Interface Account no longer has parameter dashboardUri
   - Interface Account no longer has parameter provisioningState